+++ Code/depot/src/ConsumeHistory.cpp
@@ -7,7 +7,7 @@ void ConsumeHistory::push_back(const Entry &entry)
   entries.push_back(entry);
 }
 
-void ConsumeHistory::push_back(int amount, Date date)
+void ConsumeHistory::push_back(double amount, Date date)
 {
   push_back(Entry{amount, date});
 }
